claude-bug-fix-ag-2961 (#2720)
## Summary

Add a default message to the content if the tool calls do not include a
message to prevent the following error:

![image](https://github.com/user-attachments/assets/336b15bb-2307-4317-988f-54409d2b4228)


## Type of change

- [x] Bug fix (#2712)
- [ ] New feature
- [ ] Breaking change
- [ ] Improvement
- [ ] Model update
- [ ] Other: _____________________

---

## Checklist

- [ ] Code complies with style guidelines
- [ ] Ran format/validation scripts (`./scripts/format.sh` and
`./scripts/validate.sh`)
- [ ] Self-review completed
- [ ] Documentation updated (comments, docstrings)
- [ ] Examples and guides: Relevant cookbook examples have been included
or updated (if applicable)
- [ ] Tested in clean environment
- [ ] Tests added/updated (if applicable)

---

## Additional Notes

Add any important context (deployment instructions, screenshots,
security considerations, etc.)
